[#97] Template system
